+++ test/libsolidity/smtCheckerTests/file_level/free_constant_1.sol
@@ -7,5 +7,4 @@ contract C {
 // ====
 // SMTEngine: all
 // ----
-// Warning 8195: (0-20): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (0-20): Model checker analysis was not possible because file level constants are not supported.
+// Warning 4984: (93-98): CHC: Overflow (resulting value larger than 2**256 - 1) happens here.\nCounterexample:\n\nx = 115792089237316195423570985008687907853269984665640564039457584007913129639894\n = 0\n\nTransaction trace:\nC.constructor()\nC.f(115792089237316195423570985008687907853269984665640564039457584007913129639894)