+++ test/libsolidity/smtCheckerTests/file_level/free_constant_2.sol
@@ -11,9 +11,3 @@ contract C {
 // SMTEngine: all
 // ----
 // Warning 2018: (188-278): Function state mutability can be restricted to pure
-// Warning 8195: (0-23): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (79-133): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (135-172): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (0-23): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (79-133): Model checker analysis was not possible because file level constants are not supported.
-// Warning 8195: (135-172): Model checker analysis was not possible because file level constants are not supported.