+++ test/libsolidity/smtCheckerTests/file_level/free_function_2.sol
@@ -6,6 +6,3 @@ contract C {
 function f() {}
 // ====
 // SMTEngine: all
-// ----
-// Warning 6660: (50-65): Model checker analysis was not possible because file level functions are not supported.
-// Warning 6660: (50-65): Model checker analysis was not possible because file level functions are not supported.