+++ test/libsolidity/smtCheckerTests/file_level/free_function_3.sol
@@ -1,6 +1,3 @@
 function f() view {}
 // ====
 // SMTEngine: all
-// ----
-// Warning 6660: (0-20): Model checker analysis was not possible because file level functions are not supported.
-// Warning 6660: (0-20): Model checker analysis was not possible because file level functions are not supported.