+++ test/libsolidity/smtCheckerTests/file_level/free_function_4.sol
@@ -5,5 +5,3 @@ function f()pure {
 // SMTEngine: all
 // ----
 // Warning 2072: (20-28): Unused local variable.
-// Warning 6660: (0-55): Model checker analysis was not possible because file level functions are not supported.
-// Warning 6660: (0-55): Model checker analysis was not possible because file level functions are not supported.