+++ test/libsolidity/smtCheckerTests/file_level/free_function_and_constant_1.sol
@@ -11,8 +11,8 @@ function fun(uint[] calldata _x, uint[] storage _y) view  returns (uint, uint[] 
 }
 // ====
 // SMTEngine: all
+// SMTIgnoreCex: yes
 // ----
-// Warning 8195: (0-20): Model checker analysis was not possible because file level constants are not supported.
-// Warning 6660: (214-328): Model checker analysis was not possible because file level functions are not supported.
-// Warning 8195: (0-20): Model checker analysis was not possible because file level constants are not supported.
-// Warning 6660: (214-328): Model checker analysis was not possible because file level functions are not supported.
+// Warning 4984: (190-202): CHC: Overflow (resulting value larger than 2**256 - 1) happens here.
+// Warning 4984: (190-206): CHC: Overflow (resulting value larger than 2**256 - 1) happens here.
+// Warning 6368: (315-320): CHC: Out of bounds access happens here.